finish lab3

+++ src/executors/delete_executor.cpp
@@ -18,6 +18,11 @@ std::shared_ptr<Record> DeleteExecutor::Next() {
   while (auto record = children_[0]->Next()) {
     // 通过 context_ 获取正确的锁，加锁失败时抛出异常
     // LAB 3 BEGIN
+    std::cout<<"delete executor"<<std::endl;
+    bool lock_flag = context_.GetLockManager().LockRow(context_.GetXid(), LockType::X, plan_->GetTableOid(),record->GetRid());
+    if(lock_flag == false){
+        throw DbException("delete lock failed!");
+    }
     table_->DeleteRecord(record->GetRid(), context_.GetXid(), true);
     count++;
   }
